requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
psycopg2-binary
langchain
langchain
langchain-core
langchain-community
langchain-groq
langchain-groq
python-dotenv
